# [monodroid] Add log timing categories